smart-walker/requirements.txt
opencv-python
numpy
pyserial
opencv-contrib-python